requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
statsmodels>=0.14.0
hmmlearn>=0.3.0
scipy>=1.11.0
matplotlib>=3.7.0
streamlit>=1.28.0
plotly>=5.18.0
PyYAML>=6.0
scikit-learn>=1.3.0
